Refactor AnnotationTextView
